- fixed compile error with 0d regions

+++ src/compiler/clcodegenerator.cpp
@@ -93,18 +93,18 @@ CLCodeGenerator::beginKernel( const std::vector<std::string>& outputs, const std
     _os << ", __global " << STRINGIFY(MATRIX_ELEMENT_T) << "* _region_" << *it;
 
   // And we'll need to provide the size of the region that we want the kernel to operate on.  (This is where the 'center' of the rule will be.)
-  for( unsigned int i = 0; i < dims; ++i )
+  for( int i = 0; i < (int)dims; ++i )
       _os << ", int dim_d" << i;
 
   // Finally, we need to provide some of the dimensions of each of the matrices we've passed in, so that we can calculate indices.
   for( std::vector<std::string>::const_iterator it = outputs.begin( ); it != outputs.end( ); ++it )
     {
-      for( unsigned int i = 0; i < dims-1; ++i )
+      for( int i = 0; i < (int)dims-1; ++i )
 	_os << ", int dim_" << *it << "_d" << i;
     }
   for( std::vector<std::string>::const_iterator it = inputs.begin( ); it != inputs.end( ); ++it )
     {
-      for( unsigned int i = 0; i < dims-1; ++i )
+      for( int i = 0; i < (int)dims-1; ++i )
 	_os << ", int dim_" << *it << "_d" << i;
     }
